+++ lib/ui/screens/visiting_screen.dart
@@ -1,14 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:places/domain/sight.dart';
-import 'package:places/ui/res/border_radiuses.dart';
 
+import 'package:places/ui/res/border_radiuses.dart';
 import 'package:places/ui/res/strings/strings.dart';
 import 'package:places/ui/res/text_styles.dart';
 import 'package:places/ui/res/colors.dart';
 
 import 'package:places/mocks.dart';
 
-import 'package:places/ui/screens/sight_card.dart';
+import 'package:places/ui/widgets/sight_card.dart';
 import 'package:places/ui/widgets/app_bottom_navigation_bar.dart';
 
 class VisitingScreen extends StatefulWidget {
